refactor hunt.py to also work for bighunt
